Extract stats player touch module
